hotfix2

+++ src/main/java/com/example/msgadminapi/dto/response/StatisticsResponseDto.java
@@ -17,7 +17,7 @@ public class StatisticsResponseDto {
     @JsonProperty
     private List<String> dayOfWeekList;
     @JsonProperty
-    private List<Long> grade;
+    private List<Integer> grade;
     @JsonProperty
     private int attend;
 }
